[TASK] Retrieve created version after writing to stream
After an event has been written to a stream, the store responds
with an URI that reflects this new version. The proper way to
retrieve the version value would be to issue another request for
the new URI and extract the data. However, to reduce the amount
of round-trips the version is directly extracted from the first
HTTP Location header.

```
Location: http://127.0.0.1:2113/streams/newstream/13
```
... reflects version `13` of `newstream`